# backend/internal/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Ward API
  description: >
    Self-hosted security-edge control plane over Caddy + Coraza. Sessions and
    machine API tokens both authenticate as bearer tokens.
  version: "1"
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /blocklist:
    get:
      operationId: listBlocklist
      summary: List IP block/allow rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Block" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createBlock
      summary: Create a rule (reconciles the edge)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlockInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Block" }
        default: { $ref: "#/components/responses/Error" }
  /blocklist/{id}:
    patch:
      operationId: updateBlock
      summary: Update an IP rule (partial; reconciles the edge)
      description: >-
        JSON Merge Patch (RFC 7396): only the fields present in the body change. `null` clears a field to its default, nested objects merge field by field, arrays replace as a whole, and anything absent keeps its current value.
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlockUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Block" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteBlock
      summary: Delete a rule (reconciles the edge)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
...
  /rate-limits:
    get:
      operationId: listRateLimits
      summary: List rate limits
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RateLimit" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createRateLimit
      summary: Create a rate limit (reconciles the edge)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RateLimitInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RateLimit" }
        default: { $ref: "#/components/responses/Error" }
  /rate-limits/{id}:
    patch:
      operationId: updateRateLimit
      summary: Update a rate limit (partial; reconciles the edge)
      description: >-
        JSON Merge Patch (RFC 7396): only the fields present in the body change. `null` clears a field to its default, nested objects merge field by field, arrays replace as a whole, and anything absent keeps its current value.
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RateLimitUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RateLimit" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteRateLimit
      summary: Delete a rate limit (reconciles the edge)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
  /geo-rules:
    get:
      operationId: listGeoRules
      summary: List geo rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GeoRule" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createGeoRule
      summary: Create a geo rule (reconciles the edge)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GeoRuleInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GeoRule" }
        default: { $ref: "#/components/responses/Error" }
  /geo-rules/{id}:
    patch:
      operationId: updateGeoRule
      summary: Update a geo rule (partial; reconciles the edge)
      description: >-
        JSON Merge Patch (RFC 7396): only the fields present in the body change. `null` clears a field to its default, nested objects merge field by field, arrays replace as a whole, and anything absent keeps its current value.
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GeoRuleUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GeoRule" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteGeoRule
      summary: Delete a geo rule (reconciles the edge)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token or machine API token.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Scope:
      type: string
      enum: [global, service]
    BlockMode:
      type: string
      enum: [block, allow]
...
    Block:
      type: object
      required: [id, scope, mode, cidr, source, created_at]
      properties:
        id: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
        mode: { $ref: "#/components/schemas/BlockMode" }
        service_id: { type: string, nullable: true }
        cidr: { type: string }
        reason: { type: string }
        source: { type: string }
        expires_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    BlockInput:
      type: object
      required: [cidr]
      properties:
        cidr: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
        mode: { $ref: "#/components/schemas/BlockMode" }
        service_id: { type: string, nullable: true }
        reason: { type: string }
        expires_at: { type: string, format: date-time, nullable: true }
    BlockUpdate:
      description: >-
        JSON Merge Patch (RFC 7396): only the fields present in the body change. `null` clears a field to its default, nested objects merge field by field, arrays replace as a whole, and anything absent keeps its current value.
      type: object
      properties:
        cidr: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
        mode: { $ref: "#/components/schemas/BlockMode" }
        service_id: { type: string, nullable: true }
        reason: { type: string }
        expires_at: { type: string, format: date-time, nullable: true }
...
    RateLimit:
      type: object
      required: [id, scope, max_events, window, created_at]
      properties:
        id: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
        service_id: { type: string, nullable: true }
        max_events: { type: integer }
        window: { type: string, description: "Go duration, e.g. 10s / 1m." }
        created_at: { type: string, format: date-time }
    RateLimitInput:
      type: object
      required: [scope, max_events, window]
      properties:
        scope: { $ref: "#/components/schemas/Scope" }
        service_id: { type: string, nullable: true }
        max_events: { type: integer }
        window: { type: string }
    RateLimitUpdate:
      description: >-
        JSON Merge Patch (RFC 7396): only the fields present in the body change. `null` clears a field to its default, nested objects merge field by field, arrays replace as a whole, and anything absent keeps its current value.
      type: object
      properties:
        scope: { $ref: "#/components/schemas/Scope" }
        service_id: { type: string, nullable: true }
        max_events: { type: integer }
        window: { type: string }
    GeoRule:
      type: object
      required: [id, scope, mode, countries, created_at]
      properties:
        id: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
        mode: { $ref: "#/components/schemas/BlockMode" }
        service_id: { type: string, nullable: true }
        countries:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
    GeoRuleInput:
      type: object
      required: [scope, countries]
      properties:
        scope: { $ref: "#/components/schemas/Scope" }
        mode: { $ref: "#/components/schemas/BlockMode" }
        service_id: { type: string, nullable: true }
        countries:
          type: array
          items: { type: string }
    GeoRuleUpdate:
      description: >-
        JSON Merge Patch (RFC 7396): only the fields present in the body change. `null` clears a field to its default, nested objects merge field by field, arrays replace as a whole, and anything absent keeps its current value.
      type: object
      properties:
        scope: { $ref: "#/components/schemas/Scope" }
        mode: { $ref: "#/components/schemas/BlockMode" }
        service_id: { type: string, nullable: true }
        countries:
          type: array
          items: { type: string }
...